targetid-fixes(05) Refactor GetDeviceName()
